Remove Bitboard class

+++ Board.hpp
@@ -30,7 +30,7 @@ public:
 private:
     void InitializeBitboards();
     void InitializeSquares();
-    int GetBitboardIndex(piece::PieceType pieceType, piece::PieceColor pieceColor);
+    size_t GetBitboardIndex(piece::PieceType pieceType, piece::PieceColor pieceColor);
 
     std::array<Bitboard, constants::NUM_PIECE_TYPES * constants::NUM_COLORS> m_bitboards{};
     std::array<Square, constants::NUM_SQUARES> m_squares{};
